@@ -35,8 +35,8 @@
     "electron-dl": "^3.2.1",
     "electron-is-dev": "^2.0.0",
     "electron-squirrel-startup": "^1.0.0",
-    "eslint-config-react-app": "7.0.0",
-    "immer": "^9.0.14",
+    "eslint-config-react-app": "7.0.1",
+    "immer": "9.0.13",
     "is-electron": "^2.2.0",
     "match-sorter": "^6.3.1",
     "moment": "^2.29.1",
@@ -53,7 +53,7 @@
     "react-moment": "^1.1.1",
     "react-rnd": "^10.3.5",
     "react-router-dom": "^6.3.0",
-    "react-scripts": "5.0.0",
+    "react-scripts": "5.0.1",
     "reactour": "^1.18.7",
     "serve": "^13.0.2",
     "sockette": "^2.0.6",
@@ -61,7 +61,7 @@
     "typescript": "^4.6.2",
     "universal-cookie": "^4.0.4",
     "use-api-polling": "^0.1.4",
-    "use-long-press": "^1.1.1",
+    "use-long-press": "^2.0.2",
     "web-vitals": "^2.1.4",
     "workbox-background-sync": "^6.5.1",
     "workbox-broadcast-update": "^6.5.1",
@@ -75,7 +75,7 @@
     "workbox-routing": "^6.5.1",
     "workbox-strategies": "^6.5.1",
     "workbox-streams": "^6.5.1",
-    "zustand": "^3.7.1"
+    "zustand": "^4.0.0-rc.1"
   },
   "scripts": {
     "start": "react-scripts start",
@@ -144,11 +144,11 @@
     "babel-eslint": "^10.1.0",
     "concurrently": "^7.0.0",
     "cross-env": "^7.0.3",
-    "electron": "^18.0.1",
-    "electron-builder": "^22.14.13",
+    "electron": "^19.0.2",
+    "electron-builder": "^23.0.3",
     "electron-devtools-installer": "git+https://github.com/MarshallOfSound/electron-devtools-installer.git",
     "eslint-config-airbnb": "^19.0.4",
-    "eslint-config-airbnb-typescript": "^16.1.2",
+    "eslint-config-airbnb-typescript": "^17.0.0",
     "eslint-config-prettier": "^8.5.0",
     "eslint-import-resolver-typescript": "^2.4.0",
     "eslint-plugin-flowtype": "^8.0.2",
@@ -159,7 +159,7 @@
     "eslint-plugin-prettier": "^4.0.0",
     "eslint-plugin-react": "^7.29.4",
     "eslint-plugin-react-hooks": "^4.4.0",
-    "gh-pages": "^3.2.0",
+    "gh-pages": "^4.0.0",
     "prettier": "^2.6.0",
     "react-git-info": "^2.0.1",
     "react-is": "^18.0.0",
