@@ -1,5 +1,6 @@
 import { useState } from 'react';
-import { MenuItem, ListItemIcon, Badge } from '@material-ui/core';
+import { MenuItem, ListItemIcon } from '@material-ui/core';
+import { Badge } from '@mui/material'
 import Tour from 'reactour';
 import { InfoRounded } from '@material-ui/icons';
 import useStore from '../../store/useStore';
@@ -21,7 +22,7 @@ const steps = [
 
 const TourIntegrations = ({ cally }) => {
   const [isTourOpen, setIsTourOpen] = useState(false);
-  const setTour = useStore((state) => state.setTour);
+  const setTour = useStore((state) => state.tours.setTour);
   const invisible = useStore((state) => state.tours.integrations);
 
   return (
