@@ -21,7 +21,7 @@ export default function AboutDialog({ className, children, startIcon, title }) {
         setOpen(false);
     };
 
-    useEffect(async () => {
+    useEffect(() => {
         async function fetchData() {
             const info = await getInfo()
             if (info && info.git_build_commit) {
