@@ -21,7 +21,7 @@ const step = ({ title = 'Title', text = 'Text', number = 0 }) => (
 
 const TourEffect = ({ schema }) => {
   const [isTourOpen, setIsTourOpen] = useState(false);
-  const setTour = useStore((state) => state.setTour);
+  const setTour = useStore((state) => state.tours.setTour);
 
   const yzSchema = schema && schema.properties &&
     Object.keys(schema.properties)
