@@ -8,8 +8,8 @@ import {
   CardContent,
   CardActions,
   CircularProgress,
-  Badge,
 } from '@material-ui/core';
+import { Badge } from '@mui/material'
 import logoCircle from '../assets/ring.png';
 import TourHome from '../components/Tours/TourHome';
 import useStore from '../store/useStore';
