@@ -1,12 +1,21 @@
 /* eslint-disable no-return-await */
-import { Ledfx } from '../../utils/api/ledfx';
+/* eslint-disable no-param-reassign */
+/* eslint-disable import/no-cycle */
+import produce from 'immer';
+import { Ledfx } from '../../api/ledfx';
 
-const storePresets = (get: any, set: any) => ({
+const storePresets = (set: any) => ({
   presets: {},
   getPresets: async (effectId: string) => {
     const resp = await Ledfx(`/api/effects/${effectId}/presets`);
     if (resp && resp.status === 'success') {
-      set({ presets: resp });
+      set(
+        produce((s: any) => {
+          s.presets = resp;
+        }),
+        false,
+        'gotPresets'
+      );
     }
   },
   addPreset: async (effectId: string, name: string) =>
