@@ -2,34 +2,35 @@
 /* eslint-disable no-param-reassign */
 /* eslint-disable import/no-cycle */
 import produce from 'immer';
-import { Ledfx } from '../../utils/api/ledfx';
+import { Ledfx } from '../../api/ledfx';
 
-const storeVirtuals = (get: any, set: any) => ({
+const storeVirtuals = (set: any) => ({
   virtuals: {},
   getVirtuals: async () => {
     const resp = await Ledfx('/api/virtuals');
     if (resp) {
-      set({ paused: resp.paused });
+      set(
+        produce((state: any) => {
+          state.paused = resp.paused;
+        }),
+        false,
+        'api/gotPausedState'
+      );
       if (resp && resp.virtuals) {
-        console.log('YOOOO', resp);
-        set({ v: resp });
         set(
           produce((state: any) => {
-            state.v.virtuals = resp;
+            state.virtuals = resp.virtuals;
           }),
           false,
-          'api/setVirtualEffect'
+          'api/gotVirtuals'
         );
-        console.log('YOOOO2', resp);
       }
     }
   },
   addVirtual: async (config: any) =>
     await Ledfx('/api/virtuals', 'POST', config),
   updateVirtual: async (virtId: string, active: boolean) =>
-    await Ledfx(`/api/virtuals/${virtId}`, 'PUT', {
-      active,
-    }),
+    await Ledfx(`/api/virtuals/${virtId}`, 'PUT', active),
   deleteVirtual: async (virtId: string) =>
     await Ledfx(`/api/virtuals/${virtId}`, 'DELETE'),
   clearVirtualEffect: async (virtId: string) =>
